requirements3.8.12.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
dnspython==1.16.0
eventlet==0.30.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.0.post0
gunicorn==20.1.0
idna==3.4
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
websocket-client==1.4.2
zope.event==4.5.0
zope.interface==5.5.1
